Save memory by reusing common strings.
Keep as an attribute of the class the list of all speakers, conferences
and tags. As `/api/filters` will be called often, we want to cache the
result.
These lists are computed from a `std::set` and are thus ordered.

In `TalksDB`, rather than keeping all Talks (with all `std::string`s),
use `std::string_view`s pointing too the above lists.